whitespace consistency (also explanation, see desc)
The most recent version of LDC2 (1.12 is working, later don't seem to be) weren't allowing this to compile. This allows it to compile on more recent versions, at the expense of removing `nothrow` from arithmetic operations on quantities when compiled using LDC (and *only* when using LDC).

This is of course due to the fact that LDC's--and *only* LDC's--implementation of insertInPlace has recently lost `nothrow`, which means insertAndSort, insertSorted etc. cascading on down can no longer be marked `nothrow`. This is... very annoying, but I personally would prefer not having nothrow over not being able to use LDC.